feat: auto-embed notes on save, split note_detail into SRP modules (Track E step 3)
embed.rs: background embed pipeline (chunk → OpenAI → LanceDB) with syslog logging for iOS.
folder_picker.rs: FolderPicker component extracted from NoteDetail.
recording_bar.rs: RecordingBar + start_transcription extracted from NoteDetail.
note_detail.rs: slim orchestrator using FolderPicker + RecordingBar.
Embed triggers on use_drop, skip if <50 chars or no API key.
Delete embeddings on note delete.
