+++ src/services/mod.rs
@@ -1,5 +1,6 @@
 pub mod ai;
 pub mod audio;
 pub mod constants;
+pub mod embed;
 pub mod transcription;
 pub mod vectordb;
